+++ libs/ui/src/lib/value/value.component.stories.ts
@@ -1,5 +1,6 @@
 import '@angular/localize/init';
-import { moduleMetadata } from '@storybook/angular';
+import { provideNoopAnimations } from '@angular/platform-browser/animations';
+import { applicationConfig, moduleMetadata } from '@storybook/angular';
 import type { Meta, StoryObj } from '@storybook/angular';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 
@@ -9,6 +10,9 @@ export default {
   title: 'Value',
   component: GfValueComponent,
   decorators: [
+    applicationConfig({
+      providers: [provideNoopAnimations()]
+    }),
     moduleMetadata({
       imports: [NgxSkeletonLoaderModule]
     })
@@ -102,4 +106,13 @@ export const Precision: Story = {
     precision: 3,
     value: 7.2534802394809285309
   }
+};
+
+export const WithCopyButton: Story = {
+  args: {
+    enableCopyToClipboardButton: true,
+    locale: 'en-US',
+    value: 1234.56
+  },
+  name: 'With Copy Button'
 };